eliminate static cast

+++ source/tessellation/Delaunay.hpp
@@ -134,7 +134,7 @@ private:
   vector<size_t> FindContainingTetras(size_t StartTetra, size_t point);
   vector<vector<int> > FindOuterPoints(vector<Edge> const& edges);
   bool IsTripleOut(size_t index) const;
-  int FindTripleLoc(facet const& f)const;
+  size_t FindTripleLoc(facet const& f)const;
   void AddFacetDuplicate(int index,vector<vector<int> > &toduplicate,vector<Edge>
 	const& edges,vector<bool> &checked)const;
   void AddOuterFacets(int tri,vector<vector<int> > &toduplicate,vector<Edge>
